feature/74 added BaseImageRenderer and extends the image handlers from there
