+++ packages/processors/image-processor/src/server/server-image-renderer.ts
@@ -5,6 +5,7 @@
 
 import { Logger } from '@ecopages/logger';
 import type { GenerateAttributesResult, ImageProps, RenderImageToString } from 'src/shared/image-renderer-provider';
+import { BaseImageRenderer } from '../shared/base-image-renderer';
 import type { ImageLayout, LayoutAttributes } from '../shared/constants';
 import { ImageUtils } from '../shared/image-utils';
 import type { ImageMap, ImageVariant } from './image-processor';
@@ -17,8 +18,10 @@ const appLogger = new Logger('[@ecopages/image-processor/server-image-renderer]'
  * It uses the provided props to generate the attributes
  * It also uses the image map provided by the server to generate the srcset and sizes attributes
  */
-export class ServerImageRenderer {
-  constructor(private imageMap: ImageMap) {}
+export class ServerImageRenderer extends BaseImageRenderer {
+  constructor(private imageMap: ImageMap) {
+    super();
+  }
 
   /**
    * It generates the attributes for the image element based on the provided props
@@ -69,145 +72,5 @@ export class ServerImageRenderer {
       alt: props.alt,
       ...(useResponsiveImage ? { srcset, sizes } : {}),
     };
-  }
-
-  /**
-   * Generates a stringified version of the attributes
-   * @param props
-   * @returns
-   */
-  stringifyImageAttributes(props: ImageProps): string {
-    const attributes = this.generateAttributes(props);
-    if (!attributes) return '';
-    return this.stringifyAttributes(attributes);
-  }
-
-  /**
-   * Generates a stringified version of the image element
-   * @param props
-   * @returns
-   */
-  renderImageToString({
-    src,
-    alt,
-    width,
-    height,
-    aspectRatio,
-    priority,
-    layout,
-    unstyled,
-    staticVariant,
-    ...rest
-  }: RenderImageToString): string {
-    const derivedAttributes = this.generateAttributes({
-      src,
-      alt,
-      width,
-      height,
-      aspectRatio,
-      priority,
-      layout,
-      unstyled,
-      staticVariant,
-    });
-    if (!derivedAttributes) return '';
-
-    const stringifiedAttributes = this.stringifyAttributes({
-      ...derivedAttributes,
-      ...rest,
-    });
-
-    return `<img ${stringifiedAttributes} />`;
-  }
-
-  private stringifyAttributes(attributes: GenerateAttributesResult): string {
-    const attributePairs: string[] = [];
-
-    for (const [key, value] of Object.entries(attributes)) {
-      if (value == null || value === '') continue;
-
-      if (typeof value === 'boolean') {
-        if (value) attributePairs.push(key);
-      } else {
-        attributePairs.push(`${key}="${value}"`);
-      }
-    }
-
-    return attributePairs.join(' ');
-  }
-
-  private getDimensionsAttributes(
-    width?: number,
-    height?: number,
-    aspectRatio?: string,
-    layout: ImageLayout = 'constrained',
-    unstyled?: boolean,
-  ): Partial<LayoutAttributes> {
-    const attributes: Partial<LayoutAttributes> = {};
-    const styles: [string, string][] = [];
-
-    if (!unstyled) {
-      styles.push(['object-fit', 'cover']);
-    }
-
-    if (aspectRatio) {
-      const ratio = this.parseAspectRatio(aspectRatio);
-      if (ratio) {
-        if (!unstyled) {
-          styles.push(['aspect-ratio', aspectRatio]);
-        }
-
-        if (width) {
-          const calculatedHeight = Math.round(width / ratio);
-          attributes.width = width;
-          attributes.height = calculatedHeight;
-          this.addLayoutStyles(styles, layout, width, calculatedHeight, unstyled);
-        } else if (height) {
-          const calculatedWidth = Math.round(height * ratio);
-          attributes.width = calculatedWidth;
-          attributes.height = height;
-          this.addLayoutStyles(styles, layout, calculatedWidth, height, unstyled);
-        } else if (!unstyled) {
-          styles.push(['width', '100%'], ['height', 'auto']);
-        }
-      }
-    } else if (width || height) {
-      attributes.width = width;
-      attributes.height = height;
-      this.addLayoutStyles(styles, layout, width, height, unstyled);
-    }
-
-    if (styles.length > 0 && !unstyled) {
-      attributes.style = this.generateStyleString(styles);
-    }
-
-    return attributes;
-  }
-
-  private parseAspectRatio(value: string | undefined): number | undefined {
-    if (!value) return undefined;
-
-    if (value.includes('/')) {
-      const [width, height] = value.split('/').map(Number);
-      return width / height;
-    }
-
-    return Number(value);
-  }
-
-  private generateStyleString(styles: [string, string][]): string {
-    return styles.map(([key, value]) => `${key}:${value}`).join(';');
-  }
-
-  private addLayoutStyles(
-    styles: [string, string][],
-    layout: ImageLayout,
-    width?: number,
-    height?: number,
-    unstyled?: boolean,
-  ): void {
-    if (unstyled) return;
-    const layoutStyles = ImageUtils.generateLayoutStyles({ layout, width, height });
-    styles.push(...layoutStyles);
   }
 }
